feat: add CopilotChat plugin to Neovim config
CopilotChat plugin has been added to the Neovim configuration. This
includes the addition of the plugin to the lazy-lock.json file and the
creation of a new configuration file for the plugin. Keymaps for the
plugin have also been set up in the configuration file. The which-key
plugin configuration has been updated to include a new category for
Copilot commands.